requirements.txt
#Required PIP Packages 


pandas
sodapy
plotly
python-dotenv
pytest
tkinter


